+++ src/main/java/com/github/niko91101/financetracker/dto/request/CreateTransactionRequest.java
@@ -3,12 +3,14 @@ package com.github.niko91101.financetracker.dto.request;
 import jakarta.validation.constraints.*;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class CreateTransactionRequest {
 
     @NotNull(message = "Сумма не может быть пустой")
     @Positive(message = "Сумма не может быть меньше нуля")
-    private Integer amount;
+    private BigDecimal amount;
 
     @Size(min = 5, max = 50, message = "Описание должно быть от 5 до 50 символов")
     @NotBlank(message = "Описание не может быть пустым")
